Extract out member card component and team data.
