+++ client/src/pages/About_WEPP.tsx
@@ -4,7 +4,7 @@ import wepp_diagram from "../assets/images/wepp_diagram_v2.png";
 import Typography from "@mui/material/Typography";
 import Button from "@mui/material/Button";
 import Divider from "@mui/material/Divider";
-import SidePanelLayout from "../components/SidePanelLayout";
+import SidePanelLayout from "../components/side-panels/SidePanelLayout";
 
 const useStyles = tss.create(({ theme }) => ({
   backButton: {
