+++ client/src/pages/Scenarios.tsx
@@ -4,7 +4,7 @@ import scenarios_diagram from "../assets/images/scenarios_diagram.png";
 import Typography from "@mui/material/Typography";
 import Button from "@mui/material/Button";
 import Divider from "@mui/material/Divider";
-import SidePanelLayout from "../components/SidePanelLayout";
+import SidePanelLayout from "../components/side-panels/SidePanelLayout";
 
 const useStyles = tss.create(({ theme }) => ({
   backButton: {
