+++ client/src/pages/Team.tsx
@@ -1,23 +1,12 @@
 import React from "react";
 import { tss } from "../utils/tss";
-import RogerLew from "../assets/images/roger_lew.png";
-import MarianaDobre from "../assets/images/mariana_dobre.png";
-import ErinBrooks from "../assets/images/erin_brooks.png";
-import SubhankarDas from "../assets/images/subhankar_das.jpeg";
-import ErinHanan from "../assets/images/erin_hanan.jpg";
-import WilliamBurke from "../assets/images/william_burke.png";
-import LawrenceAlawode from "../assets/images/lawrence_alawode.png";
-import MingliangLiu from "../assets/images/mingliang_liu.png";
-import JuliePadowski from "../assets/images/julie_padowski.png";
-import JennyAdam from "../assets/images/jenny_adam.png";
-import KevinBladon from "../assets/images/kevin_bladon.png";
-import RyanCole from "../assets/images/ryan_cole.png";
-import PeteRobichaud from "../assets/images/pete_robichaud.png";
 import Typography from "@mui/material/Typography";
+import MemberCard from "../components/MemberCard";
+import { groupedMembers } from "../data/team";
 import Button from "@mui/material/Button";
 import Link from "@mui/material/Link";
 import Stack from "@mui/material/Stack";
-import SidePanelLayout from "../components/SidePanelLayout";
+import SidePanelLayout from "../components/side-panels/SidePanelLayout";
 
 const useStyles = tss.create(({ theme }) => ({
   navButtons: {
@@ -58,31 +47,6 @@ const useStyles = tss.create(({ theme }) => ({
     fontSize: "2rem",
     textAlign: "center",
   },
-  memberCard: {
-    display: "inline-block",
-    padding: 10,
-    margin: 10,
-    width: 200,
-    verticalAlign: "top",
-    textAlign: "left",
-  },
-  memberImage: {
-    display: "block",
-    height: 190,
-    width: "100%",
-    objectFit: "cover" as const,
-    objectPosition: "center",
-  },
-  memberName: {
-    fontWeight: 800,
-    paddingTop: 10,
-    marginBottom: 10,
-  },
-  memberLink: {
-    color: theme.palette.accent.main,
-    fontWeight: 800,
-    textDecoration: "none",
-  },
   partnerSection: {
     padding: "40px 40px 0 40px",
     scrollMarginTop: 60,
@@ -99,148 +63,6 @@ const scrollToUniversity = (e: React.MouseEvent, id: string) => {
   document
     .getElementById(id)
     ?.scrollIntoView({ behavior: "smooth", block: "start" });
-};
-
-interface TeamMember {
-  name: string;
-  img: string;
-  role: string;
-  web: string;
-}
-
-interface UniversityGroup {
-  name: string;
-  members: TeamMember[];
-}
-
-const groupedMembers: UniversityGroup[] = [
-  {
-    name: "Washington State University",
-    members: [
-      {
-        name: "Mingliang Liu",
-        img: MingliangLiu,
-        role: "Assistant Research Professor, Civil & Environmental Engineering",
-        web: "https://ce.wsu.edu/faculty/liu-mingliang/",
-      },
-      {
-        name: "Julie Padowski",
-        img: JuliePadowski,
-        role: "Research Associate Professor, School of the Environment",
-        web: "https://environment.wsu.edu/faculty/wsu-profile/julie.padowski/",
-      },
-      {
-        name: "Jenny Adam",
-        img: JennyAdam,
-        role: "Professor, Civil & Environmental Engineering",
-        web: "https://ce.wsu.edu/faculty/adam/",
-      },
-    ],
-  },
-  {
-    name: "University of Idaho",
-    members: [
-      {
-        name: "Roger Lew",
-        img: RogerLew,
-        role: "Research Associate Professor, Virtual Technology and Design",
-        web: "https://www.uidaho.edu/people/rogerlew",
-      },
-      {
-        name: "Mariana Dobre",
-        img: MarianaDobre,
-        role: "Assistant Professor, Soil and Water Systems",
-        web: "https://www.uidaho.edu/people/mdobre",
-      },
-      {
-        name: "Erin Brooks",
-        img: ErinBrooks,
-        role: "Professor, Soil and Water Systems",
-        web: "https://www.uidaho.edu/people/ebrooks",
-      },
-      {
-        name: "Subhankar Das",
-        img: SubhankarDas,
-        role: "Postdoctoral Fellow, Soil and Water Systems",
-        web: "https://scholar.google.com/citations?user=K2ZoamkAAAAJ&hl=en",
-      },
-    ],
-  },
-  {
-    name: "University of Nevada, Reno",
-    members: [
-      {
-        name: "Erin Hanan",
-        img: ErinHanan,
-        role: "Associate Professor, Fire & Ecosystem Ecology",
-        web: "https://www.unr.edu/nres/people/hanan-erin",
-      },
-      {
-        name: "William Burke",
-        img: WilliamBurke,
-        role: "Ecohydrologic Researcher",
-        web: "https://www.wdburke.com/",
-      },
-      {
-        name: "Lawrence Alawode",
-        img: LawrenceAlawode,
-        role: "Doctoral Student, Hydrologic Sciences",
-        web: "https://www.unr.edu/hydrologic-sciences/people/students/lawrence-gbenga-alawode",
-      },
-    ],
-  },
-  {
-    name: "Oregon State University",
-    members: [
-      {
-        name: "Kevin Bladon",
-        img: KevinBladon,
-        role: "Faculty, Forest Ecosystems & Society",
-        web: "https://directory.forestry.oregonstate.edu/people/bladon-kevin",
-      },
-      {
-        name: "Ryan Cole",
-        img: RyanCole,
-        role: "Faculty Research Assistant, Wildfire and Water Security",
-        web: "https://www.researchgate.net/profile/Ryan-Cole-9",
-      },
-    ],
-  },
-  {
-    name: "US Forest Service, Rocky Mtn Research Station",
-    members: [
-      {
-        name: "Pete Robichaud",
-        img: PeteRobichaud,
-        role: "Research Engineer, Erosion Modeling and Mitigation, Wildfires",
-        web: "https://scholar.google.com/citations?user=wy3ols4AAAAJ&hl=en",
-      },
-    ],
-  },
-];
-
-const MemberCard = ({ person }: { person: TeamMember }) => {
-  const { classes } = useStyles();
-  return (
-    <div className={classes.memberCard}>
-      <img src={person.img} alt={person.name} className={classes.memberImage} />
-      <Typography
-        component="p"
-        variant="subtitle2"
-        className={classes.memberName}
-      >
-        <a
-          href={person.web}
-          target="_blank"
-          rel="noopener noreferrer"
-          className={classes.memberLink}
-        >
-          {person.name}
-        </a>
-      </Typography>
-      <Typography variant="body2">{person.role}</Typography>
-    </div>
-  );
 };
 
 const SidePanel = () => {
@@ -269,7 +91,7 @@ const SidePanel = () => {
         </Button>
         <Button
           className={classes.navButton}
-          onClick={(e) => scrollToUniversity(e, "University_of_Nevada,_Reno")}
+          onClick={(e) => scrollToUniversity(e, "University_of_Nevada_Reno")}
         >
           University of Nevada, Reno
         </Button>
@@ -284,7 +106,7 @@ const SidePanel = () => {
           onClick={(e) =>
             scrollToUniversity(
               e,
-              "US_Forest_Service,_Rocky_Mtn_Research_Station",
+              "US_Forest_Service_Rocky_Mtn_Research_Station",
             )
           }
         >
@@ -309,7 +131,7 @@ const Content = () => {
         {groupedMembers.map(({ name: university, members }) => (
           <section
             key={university}
-            id={university.replace(/\s+/g, "_")}
+            id={university.replace(/[^a-zA-Z0-9]+/g, "_")}
             className={classes.universitySection}
           >
             <Typography component="h2" className={classes.univHeading}>
